Fix Top Hits Incompatible Field Types on Sort Across Indices (#142046)
* addresses top hits incompatible field types on sort across indices
